feat: add get, update, and delete user update
- need to make tests

+++ utils/verify_userid.go
@@ -5,10 +5,10 @@ import (
 	"strconv"
 )
 
-func VerifyUserId(UserID string) (uint64, error) {
+func VerifyUserId(UserID string) (uint, error) {
 	userID, err := strconv.ParseUint(UserID, 10, 64)
 	if err != nil {
 		return 0, errors.New("cannot parse id")
 	}
-	return userID, nil
+	return uint(userID), nil
 }
